Various configuration changes.
After using this thing in anger there's a few issues which these changes are designed to address.

Installing services via cloud-init packages seems unreliable - so instead we add them to the cloud init scripting section.

We set up a password for each user on the system (matching their username).  This sounds insecure but we also disallow password authentication over SSH - so this password is only usable if you have physical access to the box (or your hosting service gives you console access).

And dokku.txt lists the stuff you need to do to get a Dokku app running.  They're basically notes for me, so I can finish implementing it all.